[PROD-14034] Reduce loki query range to avoid rounding errors
We think that passing the exact max query range to loki can lead to some
rounding error if the values do not align with the current moon phase.
Reducing drastically (from the default 30d1h default value) the query
range should limit this issue.
Anyway, loki should not be used like that from the api, it's better
suited for grafana and we should think going back to archiving logs in
argo workflows directly.